+++ src/client/components/simple-list/simple-list.scss
@@ -1,67 +1,42 @@
 @import '../../imports';
 
 .simple-list {
-  width: 100%;
-  height: 100%;
+  overflow: auto;
 
-  .scroller {
-    height: 100%;
+  .row {
+    display: flex;
+    padding: 10px;
 
-    .column-headers {
-      font-size: 1.2em;
-      border-bottom: 1px solid $border-light;
-      height: 100%;
+    &:not(:first-child) {
+      border-top: 1px solid $border-light;
+    }
 
-      background-color: $white;
+    > * {
 
-      > * {
-        display: inline-block;
+    }
+
+    .svg-icon {
+      width: 20px;
+      height: 20px;
+    }
+
+    .text {
+      flex: 1 1;
+      padding: 0 10px;
+
+      .title {
+        font-size: 1.2em;
+        color: $text-medium;
       }
 
-      .sort-arrow {
-        width: 10px;
-        margin-left: 10px;
-
-        &.ascending {
-          transform: rotate(180deg);
-        }
+      .description {
+        color: $text-lighter;
       }
     }
 
-    .top-right-corner {
-      border-bottom: 1px solid $border-light;
-      background-color: $white;
-    }
-
-    .row {
-      padding: 12px 0 5px 0;
-      border-bottom: 1px solid $border-light;
-      white-space: nowrap;
-
-      > * {
-        display: inline-block;
-      }
-
-      .cell {
-        &.has-icon {
-          padding-left: 25px;
-        }
-
-        .icon {
-          position: absolute;
-          top: -3px;
-          left: 0;
-          width: 20px;
-        }
-      }
-
-      &.action {
-        cursor: pointer;
-
-        .icon {
-          width: 20px;
-          height: 20px;
-        }
+    .actions {
+      button {
+        @extend %button-light;
       }
     }
   }
